Jar for DigitalOcean

+++ src/main/java/com/schindig/utils/Parameters.java
@@ -41,6 +41,6 @@ public class Parameters {
 
     public Integer listID;
 
-    public Double amount;
+    public Integer amount;
 
 }
